# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ThetaFrame API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vision-frames:
    get:
      operationId: getVisionFrameCollection
      tags: [vision-frames]
      summary: Get the current user's vision frame (collection alias)
      responses:
        "200":
          description: Vision frame
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VisionFrame"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createVisionFrame
      tags: [vision-frames]
      summary: Create or update the user's vision frame
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertVisionFrameBody"
      responses:
        "200":
          description: Created or updated vision frame
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VisionFrame"
  /vision-frames/me:
    get:
      operationId: getVisionFrame
      tags: [vision-frames]
      summary: Get the user's vision frame
      responses:
        "200":
          description: Vision frame
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VisionFrame"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: upsertVisionFrame
      tags: [vision-frames]
      summary: Create or update the user's vision frame
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertVisionFrameBody"
      responses:
        "200":
          description: Upserted vision frame
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VisionFrame"
...
  /mobile/quick-capture:
    post:
      operationId: createMobileQuickCapture
      tags: [mobile]
      summary: Capture a mobile shortcut item into the correct live lane
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MobileQuickCaptureBody"
      responses:
        "200":
          description: Captured item written to the target lane
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MobileQuickCaptureResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /mobile/devices/register:
    post:
      operationId: registerMobileDevice
      tags: [mobile]
      summary: Register or refresh a simulated mobile device
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterMobileDeviceBody"
      responses:
        "200":
          description: Registered mobile device
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MobileDevice"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /mobile/devices/{id}/deactivate:
    post:
      operationId: deactivateMobileDevice
      tags: [mobile]
      summary: Deactivate a simulated mobile device
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated mobile device
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MobileDevice"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /mobile/notifications/{id}/mark-sent:
    post:
      operationId: markMobileNotificationOutboxItemSent
      tags: [mobile]
      summary: Mark a mobile notification outbox item as sent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated mobile notification outbox item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MobileNotificationOutboxItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /mobile/notifications/{id}/simulate-dispatch:
    post:
      operationId: simulateDispatchMobileNotificationOutboxItem
      tags: [mobile]
      summary: Simulate dispatching a queued mobile notification to an active device
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SimulateDispatchMobileNotificationBody"
      responses:
        "200":
          description: Updated mobile notification outbox item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MobileNotificationOutboxItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Dispatch is not possible in the current state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /mobile/notifications/{id}/acknowledge:
    post:
      operationId: acknowledgeMobileNotificationOutboxItem
      tags: [mobile]
      summary: Acknowledge a mobile notification outbox item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated mobile notification outbox item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MobileNotificationOutboxItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    TierTask:
      type: object
      properties:
        id:
          type: string
        text:
          type: string
        completed:
          type: boolean
        emoji:
          type: ["string", "null"]
      required:
        - id
        - text
        - completed
    TimeBlock:
      type: object
      properties:
        id:
          type: string
        startTime:
          type: string
          description: HH:MM format
        action:
          type: string
      required:
        - id
        - startTime
        - action
    DailyFrame:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        date:
          type: string
          description: YYYY-MM-DD
        colourState:
          type: string
          enum: [green, yellow, red, blue, purple]
        tierA:
          type: array
          items:
            $ref: "#/components/schemas/TierTask"
        tierB:
          type: array
          items:
            $ref: "#/components/schemas/TierTask"
        timeBlocks:
          type: array
          items:
            $ref: "#/components/schemas/TimeBlock"
        microWin:
          type: ["string", "null"]
        skipProtocolUsed:
          type: boolean
        skipProtocolChoice:
          type: ["string", "null"]
          enum: ["micro-win", "intentional-recovery", null]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - date
        - colourState
        - tierA
        - tierB
        - timeBlocks
        - skipProtocolUsed
...
    VisionGoal:
      type: object
      properties:
        id:
          type: string
        text:
          type: string
        emoji:
          type: ["string", "null"]
      required:
        - id
        - text
    VisionFrame:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        goals:
          type: array
          items:
            $ref: "#/components/schemas/VisionGoal"
        nextSteps:
          type: array
          items:
            $ref: "#/components/schemas/VisionGoal"
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - goals
        - nextSteps
    UpsertVisionFrameBody:
      type: object
      properties:
        goals:
          type: array
          items:
            $ref: "#/components/schemas/VisionGoal"
        nextSteps:
          type: array
          items:
            $ref: "#/components/schemas/VisionGoal"
      required:
        - goals
        - nextSteps
...
    MobileNotificationOutboxDeliveryStatus:
      type: string
      enum: [queued, sent, cancelled, acknowledged]
    MobileDevicePlatform:
      type: string
      enum: [ios, android]
    MobileDeviceDeliveryProvider:
      type: string
      enum: [ios_local_notification, android_notification]
    MobileQuickCaptureIntent:
      type: string
      enum: [current_work]
    MobileQuickCaptureChannel:
      type: string
      enum: [ios_shortcut, android_shortcut]
    MobileQuickCaptureBody:
      type: object
      properties:
        intent:
          $ref: "#/components/schemas/MobileQuickCaptureIntent"
        text:
          type: string
        captureChannel:
          $ref: "#/components/schemas/MobileQuickCaptureChannel"
      required:
        - intent
        - text
        - captureChannel
    MobileQuickCaptureResponse:
      type: object
      properties:
        lane:
          type: string
          enum: [daily]
        route:
          type: string
        deepLink:
          type: string
        dailyFrame:
          $ref: "#/components/schemas/DailyFrame"
        capturedTaskId:
          type: string
      required:
        - lane
        - route
        - deepLink
        - dailyFrame
        - capturedTaskId
    MobileDevice:
      type: object
      properties:
        id:
          type: integer
        installationId:
          type: string
        deviceLabel:
          type: string
        platform:
          $ref: "#/components/schemas/MobileDevicePlatform"
        deliveryProvider:
          $ref: "#/components/schemas/MobileDeviceDeliveryProvider"
        isActive:
          type: boolean
        lastSeenAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - installationId
        - deviceLabel
        - platform
        - deliveryProvider
        - isActive
        - lastSeenAt
        - createdAt
        - updatedAt
...
    RegisterMobileDeviceBody:
      type: object
      properties:
        installationId:
          type: string
        deviceLabel:
          type: string
        platform:
          $ref: "#/components/schemas/MobileDevicePlatform"
      required:
        - installationId
        - deviceLabel
        - platform
    SimulateDispatchMobileNotificationBody:
      type: object
      properties:
        deviceId:
          type: integer
    MobileNotificationOutboxItem:
      type: object
      properties:
        id:
          type: integer
        sourceLane:
          type: string
        sourceEventId:
          type: integer
        notificationCategory:
          type: string
          enum: [life_ledger_due]
        route:
          type: string
        deepLink:
          type: string
        eventName:
          type: string
        executionDate:
          type: ["string", "null"]
        sourceType:
          type: ["string", "null"]
        reminderState:
          type: string
          enum: [scheduled, snoozed]
        reminderLeadDays:
          type: array
          items:
            type: integer
        scheduledFor:
          type: string
          format: date-time
        deliveryStatus:
          $ref: "#/components/schemas/MobileNotificationOutboxDeliveryStatus"
        deliveredDeviceId:
          type: ["integer", "null"]
        deliveredDeviceLabel:
          type: ["string", "null"]
        deliveryProvider:
          oneOf:
            - $ref: "#/components/schemas/MobileDeviceDeliveryProvider"
            - type: "null"
        sentAt:
          type: ["string", "null"]
          format: date-time
        acknowledgedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - sourceLane
        - sourceEventId
        - notificationCategory
        - route
        - deepLink
        - eventName
        - reminderState
        - reminderLeadDays
        - scheduledFor
        - deliveryStatus
        - createdAt
        - updatedAt
...